Tutorial Order Change
You now place two Pawns and move two Pawns during the Tutorial because
there is no way to position them in such a way that there are no valid
options (and thus the Tutorial is broken).

Pieces could be placed in such a way that a second Bishop can't be
placed and moved in such a way that the Knight would have nowhere to go.